fixed usSlots and fetchSlots
