+++ src/hooks/useSlots.ts
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import { fetchSlots } from "@/api/appointmentApi";
 import { AvailabilityType } from "@/types/type";
+import { generateAvailabilities } from "@/lib/functions/helpers";
 
 export function useSlots(practitionerId: string | null, currentStart: Date) {
   const [availabilities, setAvailabilities] = useState<AvailabilityType[][]>([]);
@@ -8,6 +9,9 @@ export function useSlots(practitionerId: string | null, currentStart: Date) {
   const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
+    const endDateObj = new Date(currentStart);
+    endDateObj.setDate(currentStart.getDate() + 6); // sur 7 jours
+    const endDate = endDateObj.toISOString().split("T")[0];
     if (!practitionerId) {
       setAvailabilities([]);
       return;
@@ -17,7 +21,15 @@ export function useSlots(practitionerId: string | null, currentStart: Date) {
     setError(null);
 
     fetchSlots(practitionerId, currentStart)
-      .then((slots) => setAvailabilities(slots))
+      .then((slots) => {
+        const genSlots = generateAvailabilities(
+          currentStart,
+          new Date(endDate),
+          slots.openings,
+          slots.blockedSlots
+        );
+        setAvailabilities(genSlots);
+      })
       .catch((err) => setError(err))
       .finally(() => setLoading(false));
   }, [practitionerId, currentStart]);
